refactor: `@hooks/useMyInfoQuery`의 커스텀 훅을 `@hooks/auth`로 이동 (#296)
