+++ client/src/hooks/auth.ts
@@ -1,8 +1,11 @@
 import type { SignOutResult } from '@apis/auth';
+import type { GetMyInfoResult } from '@apis/user';
 import type { UseMutationOptions } from '@tanstack/react-query';
+import type { AxiosError } from 'axios';
 
 import { signOut } from '@apis/auth';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { getMyInfo } from '@apis/user';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
 import queryKeyCreator from '@/queryKeyCreator';
 
@@ -17,9 +20,28 @@ export const useSignOutMutation = (
   return mutation;
 };
 
+/* ============================ [ myInfo ] ================================ */
+// 로그인한 사용자의 정보
+
+export const useMyInfoQuery = () => {
+  const key = queryKeyCreator.me();
+  const query = useQuery<GetMyInfoResult, AxiosError>(key, getMyInfo);
+
+  return query;
+};
+
+export const useMyInfoQueryData = () => {
+  const queryClient = useQueryClient();
+  const key = queryKeyCreator.me();
+  const me = queryClient.getQueryData<GetMyInfoResult>(key);
+
+  return me;
+};
+
 export const useSetMyInfoQueryData = () => {
   const key = queryKeyCreator.me();
   const queryClient = useQueryClient();
+
   const removeMyInfoQueryData = () => {
     queryClient.setQueryData(key, () => null);
   };
